makunbound: use ECL_SETQ and not ECL_SET
The specification says that it makes the symbol unbound; that paired with the
definition of BOUNDP indicates, that what is meant is the binding in the dynamic
environment (and for consistency we want to make it behave as a pair to BOUNDP).

Note that this behavioral change matches implementations like CCL and SBCL, so
this change improves the compatibility.

The issue with interpretation comes from the subheader "Side Effects" in the
standard that states: "The value cell of symbol is modified", it should state
that "The value cell of symbol might be modified".